refactor(operators): share race-arm bookkeeping
- Extract duplicated winner-election and loser-disposal logic from
  RaceWitness<T> and RaceCoordinator<T> into a single RaceArms<T> helper
- Both coordinators now delegate registration, win, and discard to the
  shared helper; behavior is unchanged
- Removes the ~50% duplicated block flagged by the SonarCloud gate
